+++ CHANGELOG.md
@@ -25,6 +25,7 @@
     * `updateIndividualAccount` query
     * `updateProfile` query
 * `deactivateBankAccount` mock mutation was removed
+* `listBeneficiaries` mock query was removed
 
 ## 1.16 - 06/13/2023